requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
requests==2.21.0
google-auth==1.6.2
google-cloud-datastore==1.7.3
